Create JumpSearch.java
Jump search is an searching algorithm that runs in O(sqrt(n)) time. It performs better than linear search but slower than binary search.
Jump search would be ideal in cases where traversing back in an array is a costly operation. A binary search could perform lot of back traversals to indexes if the element is present at the beginning or if the searched element is even smaller than the starting element.
Hence a jump search would only require one jump as compared to binary search that would take O(log(n)) jumps in the worst case